+++ src/components/LetterAccordion/styles.ts
@@ -10,11 +10,10 @@ const style = {
     gap: 0.5rem;
     overflow-y: hidden;
     height: 100%;
-    padding-inline: 0.625rem;
   `,
   contentText: (isOpen: boolean, line: number) => css`
     position: relative;
-    height: ${isOpen ? `17.2rem` : `${line * 1.225}rem`};
+    height: ${isOpen ? `14.8rem` : `${line * 1.225}rem`};
     color: var(--kakao-logo, #000);
     font-weight: 400;
     font-style: normal;
@@ -26,7 +25,6 @@ const style = {
     position: absolute;
     top: 0;
     left: 0;
-    z-index: 1;
   `,
   originalText: (line: number) => css`
     position: absolute;
