+++ src/pages/LetterReceptionPage/ReplyToLetter/index.tsx
@@ -10,8 +10,8 @@ import useBoolean from '@/hooks/useBoolean';
 import LetterLengthDate from '@/components/LetterLengthDate';
 import LetterHeader from '@/components/LetterHeader';
 import Chip from '@/components/Chip';
-import ReceptionPolaroid from '../components/ReceptionPolaroid';
 import LetterContent from '../components/LetterContent';
+import ReceptionPolaroid from '../components/ReceptionPolaroid';
 
 interface ReplyToLetterProps {
   onPrev: () => void;
@@ -44,12 +44,13 @@ const ReplyToLetter = ({ onPrev }: ReplyToLetterProps) => {
             id="1"
             text="여기까지가 사시오 진정 행복하길 바라겠소 이 맘만 가져가오 기나긴 그대 침묵을 이별로 받아두겠소 행여 이 맘 다칠까 근심은 접어두오 오 사랑한 사람이여 더 이상 못보아도 사실 그대 있음으로 힘겨운 날들을 견뎌 왔음에 감사하오 좋은 사람 만나오 사는 동안 날 잊고 사시오 진정 행복하길 바라겠소 이 맘만 가져가오여기까지가 사시오 진정 행복하길 바라겠소 이 맘만 가져가오 기나긴 그대 침묵을 이별로 받아두겠소 행여 이 맘 다칠까 근심은 접어두오 오 사랑한 사람이여 더 이상 못보아도 사실 그대 있음으로 힘겨운 날들을 견뎌 왔음에 감사하오 좋"
             date={new Date()}
+            nickname="낯선 고양이"
             line={2}
             type="send"
           />
           <ReceptionPolaroid />
         </LetterCard>
-        <LetterCard isOpen={true}>
+        <LetterCard isOpen={true} css={{ marginBottom: '1rem' }}>
           <LetterHeader fromOrTo="To" nickname="낯선 고양이" />
           <LetterTextarea
             name="content"
